backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
python-dotenv==1.0.1
pydantic==2.8.2
openai==1.43.0
httpx==0.27.2

